Added the CRUD for promotion management system

+++ Human Resources/Human Resources/Controllers/PromotionController.cs
@@ -11,7 +11,7 @@ namespace Human_Resources.Controllers
         private readonly IPromotionService _service;
         private readonly IEmployeeService _empService;
         private readonly ILogger<PromotionController> _logger;
-        public PromotionController(IPromotionService service, IEmployeeService empService, ILogger<PromotionController> logger) 
+        public PromotionController(IPromotionService service, IEmployeeService empService, ILogger<PromotionController> logger)
         {
             _service = service;
             _empService = empService;
@@ -38,7 +38,7 @@ namespace Human_Resources.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> AddPromotion(PromotionViewModel promotionVM)
         {
-            if(!ModelState.IsValid)
+            if (!ModelState.IsValid)
             {
                 var Positiondropdowns = await _service.GetPositiondropdowns();
                 var Employeedropdowns = await _service.GetEmployeedropdowns();
@@ -73,10 +73,10 @@ namespace Human_Resources.Controllers
                         EducationalLevel = employee.EducationalLevel
 
                     };
-                    
+
                     var Positions = await _service.GetPositiondropdowns();
-                    
-                    
+
+
                     ViewBag.Positions = new SelectList(Positions.Positions, "Id", "PositionName");
                     await _empService.UpdateEmployee(EmployeeVm);
 
@@ -89,7 +89,7 @@ namespace Human_Resources.Controllers
             {
                 return View("The selected Employee doesn't exist");
 
-            }    
+            }
         }
 
         [HttpGet]
@@ -110,10 +110,10 @@ namespace Human_Resources.Controllers
                 };
                 var Positiondropdowns = await _service.GetPositiondropdowns();
                 var Employeedropdowns = await _service.GetEmployeedropdowns();
-                ViewBag.Positions = new SelectList(Positiondropdowns.Positions, "Id", "PositionName");
-                ViewBag.Employeedropdowns = new SelectList(Employeedropdowns.Employees, "Id", "EmployeeName");
+                ViewBag.PositionFrom = new SelectList(Positiondropdowns.Positions, "Id", "PositionName");
+                ViewBag.PositionTo = new SelectList(Positiondropdowns.Positions, "Id", "PositionName");
+                ViewBag.Employeedropdowns = new SelectList(Employeedropdowns.Employees, "Id", "Name");
                 return View(promotionVm);
-
             }
             else
             {
@@ -123,18 +123,53 @@ namespace Human_Resources.Controllers
         [HttpPost]
         public async Task<IActionResult> EditPromotion(PromotionViewModel promotionVm)
         {
-            if(!ModelState.IsValid)
+            if (!ModelState.IsValid)
             {
                 var Positiondropdowns = await _service.GetPositiondropdowns();
                 var Employeedropdowns = await _service.GetEmployeedropdowns();
-                ViewBag.Positions = new SelectList(Positiondropdowns.Positions, "Id", "PositionName");
-                ViewBag.Employeedropdowns = new SelectList(Employeedropdowns.Employees, "Id", "EmployeeName");
+                ViewBag.PositionFrom = new SelectList(Positiondropdowns.Positions, "Id", "PositionName");
+                ViewBag.PositionTo = new SelectList(Positiondropdowns.Positions, "Id", "PositionName");
+                ViewBag.Employeedropdowns = new SelectList(Employeedropdowns.Employees, "Id", "Name",promotionVm.EmployeeId);
                 return View(promotionVm);
             }
             else
             {
-                await _service.UpdatePromotion(promotionVm);
-                return RedirectToAction("Index", "Promotion");
+                _logger.LogInformation($"The number is {promotionVm.EmployeeId}");
+                var employee = await _empService.GetById(promotionVm.EmployeeId);
+                if (employee != null)
+                {
+                    employee.PositionId = promotionVm.toPositionId;
+                    using (var stream = new FileStream("wwwroot/images/" + employee.PhotoURL, FileMode.Open))
+                    {
+                        var file = new FormFile(stream, 0, stream.Length, null, Path.GetFileName(employee.PhotoURL));
+                        var EmployeeVm = new EmployeeViewModel()
+                        {
+                            Id = employee.Id,
+                            DepartmentId = employee.DepartmentId,
+                            PositionId = employee.PositionId,
+                            Sex = employee.Sex,
+                            Name = employee.Name,
+                            PhotoURL = file,
+                            EducationalFieldId = employee.EducationalFieldId,
+                            Email = employee.Email,
+                            EducationalLevel = employee.EducationalLevel
+
+                        };
+
+                        var Positions = await _service.GetPositiondropdowns();
+
+
+                        ViewBag.Positions = new SelectList(Positions.Positions, "Id", "PositionName");
+                        await _empService.UpdateEmployee(EmployeeVm);
+
+                    }
+                    await _service.UpdatePromotion(promotionVm);
+                    return RedirectToAction("Index", "Promotion");
+                }
+                else
+                {
+                    return View("The employee Returned doesn't exist");
+                }
             }
 
         }
@@ -143,11 +178,21 @@ namespace Human_Resources.Controllers
             var promotion = await _service.GetById(id);
             if (promotion != null)
             {
+                PromotionViewModel promotionVm = new PromotionViewModel()
+                {
+                    Id = promotion.Id,
+                    Reason = promotion.Reason,
+                    fromPositionId = promotion.fromPositionId,
+                    toPositionId = promotion.toPositionId,
+                    PositionChange = promotion.PositionChange,
+                    EmployeeId = promotion.EmployeeId,
+                };
                 var Positiondropdowns = await _service.GetPositiondropdowns();
                 var Employeedropdowns = await _service.GetEmployeedropdowns();
-                ViewBag.Positions = new SelectList(Positiondropdowns.Positions, "Id", "PositionName");
-                ViewBag.Employeedropdowns = new SelectList(Employeedropdowns.Employees, "Id", "EmployeeName");
-                return View(promotion);
+                ViewBag.PositionFrom = new SelectList(Positiondropdowns.Positions, "Id", "PositionName");
+                ViewBag.PositionTo = new SelectList(Positiondropdowns.Positions, "Id", "PositionName");
+                ViewBag.Employeedropdowns = new SelectList(Employeedropdowns.Employees, "Id", "Name",promotion.EmployeeId);
+                return View(promotionVm);
             }
             else
             {
@@ -170,18 +215,49 @@ namespace Human_Resources.Controllers
                     fromPositionId = promotion.fromPositionId,
                     toPositionId = promotion.toPositionId
 
-
                 };
-                await _service.DeletePromotion(PromotionVm);
-                return RedirectToAction("Index", "Promotion");
+                var employee = await _empService.GetById(PromotionVm.EmployeeId);
+                if (employee != null)
+                {
+                    employee.PositionId = PromotionVm.fromPositionId;
+                    using (var stream = new FileStream("wwwroot/images/" + employee.PhotoURL, FileMode.Open))
+                    {
+                        var file = new FormFile(stream, 0, stream.Length, null, Path.GetFileName(employee.PhotoURL));
+                        var EmployeeVm = new EmployeeViewModel()
+                        {
+                            Id = employee.Id,
+                            DepartmentId = employee.DepartmentId,
+                            PositionId = employee.PositionId,
+                            Sex = employee.Sex,
+                            Name = employee.Name,
+                            PhotoURL = file,
+                            EducationalFieldId = employee.EducationalFieldId,
+                            Email = employee.Email,
+                            EducationalLevel = employee.EducationalLevel
+
+                        };
+
+                        var Positions = await _service.GetPositiondropdowns();
+
+
+                        ViewBag.Positions = new SelectList(Positions.Positions, "Id", "PositionName");
+                        await _empService.UpdateEmployee(EmployeeVm);
+
+                    }
+                    await _service.DeletePromotion(PromotionVm);
+                    return RedirectToAction("Index", "Promotion");
+                }
+                else
+                {
+                    return View("The object doesn't exist");
+                }
+
 
             }
             else
             {
                 return View("The object doesn't exist");
             }
-
-
         }
     }
 }
